only:int rpm -> float rpm

+++ firmware/controllers/algo/antilag_system.cpp
@@ -32,11 +32,11 @@ bool AntilagSystemBase::isInsideALSSwitchCondition() {
 	}
 }
 
-bool AntilagSystemBase::isALSMinRPMCondition(int rpm) const {
+bool AntilagSystemBase::isALSMinRPMCondition(float rpm) const {
 	return engineConfiguration->ALSMinRPM < rpm;
 }
 
-bool AntilagSystemBase::isALSMaxRPMCondition(int rpm) const {
+bool AntilagSystemBase::isALSMaxRPMCondition(float rpm) const {
 	return engineConfiguration->ALSMaxRPM > rpm;
 }
 
@@ -64,7 +64,7 @@ bool AntilagSystemBase::isInsideALSTimerCondition() {
 	return ALStime < engineConfiguration->ALSMaxDuration;
 }
 
-bool AntilagSystemBase::isAntilagConditionMet(int rpm) {
+bool AntilagSystemBase::isAntilagConditionMet(float rpm) {
 
 
 	ALSMinRPMCondition = isALSMinRPMCondition(rpm);
@@ -89,7 +89,7 @@ todo: looking for a hero to figure out unit test part of this
 }
 
 void AntilagSystemBase::update() {
-	int rpm = Sensor::getOrZero(SensorType::Rpm);
+	float rpm = Sensor::getOrZero(SensorType::Rpm);
     isAntilagCondition = engineConfiguration->antiLagEnabled && isAntilagConditionMet(rpm);
 
 	if (!ALSMaxRPMCondition) {
