+++ firmware/controllers/algo/antilag_system.h
@@ -16,15 +16,15 @@ class AntilagSystemBase : public antilag_system_state_s {
 public:
 	void update();
 
-    bool isALSMinRPMCondition(int rpm) const;
-	bool isALSMaxRPMCondition(int rpm) const;
+	bool isALSMinRPMCondition(float rpm) const;
+	bool isALSMaxRPMCondition(float rpm) const;
 	bool isALSMinCLTCondition() const;
 	bool isALSMaxCLTCondition() const;
 	bool isALSMaxThrottleIntentCondition() const;
 	bool isInsideALSSwitchCondition();
 	bool isInsideALSTimerCondition();
     /* enabled and all conditions above */
-	bool isAntilagConditionMet(int rpm);
+	bool isAntilagConditionMet(float rpm);
 
 private:
 	Timer ALStimer;
