+++ firmware/controllers/algo/engine.cpp
@@ -222,7 +222,7 @@ void Engine::updateSlowSensors() {
 	updateSwitchInputs();
 
 #if EFI_SHAFT_POSITION_INPUT
-	int rpm = Sensor::getOrZero(SensorType::Rpm);
+	float rpm = Sensor::getOrZero(SensorType::Rpm);
 	triggerCentral.isEngineSnifferEnabled = rpm < engineConfiguration->engineSnifferRpmThreshold;
 	getEngineState()->sensorChartMode = rpm < engineConfiguration->sensorSnifferRpmThreshold ? engineConfiguration->sensorChartMode : SC_OFF;
